+++ h2/src/main/org/h2/command/ddl/DropDomain.java
@@ -44,7 +44,7 @@ public class DropDomain extends SchemaCommand {
     }
 
     @Override
-    public int update() {
+    public long update() {
         session.getUser().checkAdmin();
         session.commit(true);
         Schema schema = getSchema();
