+++ h2/src/main/org/h2/command/ddl/DropSchema.java
@@ -36,7 +36,7 @@ public class DropSchema extends DefineCommand {
     }
 
     @Override
-    public int update() {
+    public long update() {
         session.getUser().checkSchemaAdmin();
         session.commit(true);
         Database db = session.getDatabase();
