+++ h2/src/main/org/h2/command/ddl/DropSynonym.java
@@ -30,7 +30,7 @@ public class DropSynonym extends SchemaCommand {
     }
 
     @Override
-    public int update() {
+    public long update() {
         session.commit(true);
         session.getUser().checkAdmin();
